AEIM-1359: remove ReleaseBuilder
This moves checkout of the source/shared/unshared config to Artifact.
The only remaining responsibility of the ReleaseBuilder was checking out
the deploy config; that's now merged into Release. AEIM-1357 will
abstract the checkout stuff.

Also renames Artifact to Fauxpaas::Artifact as expected
